refactor(markov): move pure helpers to MarkovModel; keep IO in cli
